requirements.txt
openpyxl==3.1.5
pdfplumber==0.11.4
reportlab==4.2.2
Pillow==10.4.0